SEQNG-272 Remove some scalastyle:off annotation spliting some methods and files
